Move no-op outbox classes out of ReceiveFeature
