+++ src/NServiceBus.Core/Pipeline/Incoming/NoOpOutbox/NoOpOutboxStorage.cs
@@ -1,0 +1,31 @@
+﻿namespace NServiceBus.Features
+{
+    using System.Threading.Tasks;
+    using Extensibility;
+    using NServiceBus.Outbox;
+
+    class NoOpOutboxStorage : IOutboxStorage
+    {
+        public Task<OutboxMessage> Get(string messageId, ContextBag options)
+        {
+            return NoOutboxMessageTask;
+        }
+
+        public Task Store(OutboxMessage message, OutboxTransaction transaction, ContextBag options)
+        {
+            return TaskEx.CompletedTask;
+        }
+
+        public Task SetAsDispatched(string messageId, ContextBag options)
+        {
+            return TaskEx.CompletedTask;
+        }
+
+        public Task<OutboxTransaction> BeginTransaction(ContextBag context)
+        {
+            return Task.FromResult<OutboxTransaction>(new NoOpOutboxTransaction());
+        }
+
+        static Task<OutboxMessage> NoOutboxMessageTask = Task.FromResult<OutboxMessage>(null);
+    }
+}
