+++ src/NServiceBus.Core/Pipeline/Incoming/NoOpOutbox/NoOpOutboxTransaction.cs
@@ -1,0 +1,17 @@
+﻿namespace NServiceBus.Features
+{
+    using System.Threading.Tasks;
+    using NServiceBus.Outbox;
+
+    class NoOpOutboxTransaction : OutboxTransaction
+    {
+        public void Dispose()
+        {
+        }
+
+        public Task Commit()
+        {
+            return TaskEx.CompletedTask;
+        }
+    }
+}
