+++ src/NServiceBus.Core/Pipeline/Incoming/NoOpOutbox/NoOpSynchronizedStorage.cs
@@ -1,0 +1,14 @@
+﻿namespace NServiceBus.Features
+{
+    using System.Threading.Tasks;
+    using Extensibility;
+    using Persistence;
+
+    class NoOpSynchronizedStorage : ISynchronizedStorage
+    {
+        public Task<CompletableSynchronizedStorageSession> OpenSession(ContextBag contextBag)
+        {
+            return NoOpSynchronizedStorageAdapter.EmptyResult;
+        }
+    }
+}
